pyproject.toml
[project]
name = "techrag"
version = "0.1.0"
description = "Local-first RAG pipeline for technical documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn[standard]==0.34.2",
  "streamlit==1.44.1",
  "langchain==0.3.23",
  "langchain-community==0.3.21",
  "langchain-ollama==0.3.2",
  "langchain-groq==0.3.2",
  "langchain-huggingface==0.1.2",
  "langchain-text-splitters==0.3.8",
  "sentence-transformers==3.4.1",
  "faiss-cpu==1.11.0",
  "rank-bm25==0.2.2",
  "pypdf==5.4.0",
  "beautifulsoup4==4.13.4",
  "pyyaml==6.0.2",
  "python-multipart==0.0.20",
  "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.uv]
package = false
